partial tests integrations

+++ src/Ativos.Communication/Requests/RequestLoginJson.cs
@@ -2,6 +2,6 @@ namespace Ativos.Communication.Requests;
 
 public class RequestLoginJson
 {
-    public int Matricula { get; set; }
+    public long Matricula { get; set; }
     public string Password { get; set; }
 }
